# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Login with phone or username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/google:
    post:
      operationId: googleAuth
      tags: [Auth]
      summary: Sign in or register with Google
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Profile ─────────────────────────────────────────────────
...
  /contest:
    post:
      operationId: createContest
      tags: [Contest]
      summary: Admin - create a contest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContestRequest'
      responses:
        '201':
          description: Contest created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contest'
  /contest/{id}:
    put:
      operationId: updateContest
      tags: [Contest]
      summary: Admin - update a contest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContestRequest'
      responses:
        '200':
          description: Contest updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contest'

  # ─── Notifications ───────────────────────────────────────────
...
components:
  schemas:
...
    RegisterRequest:
      type: object
      required: [username, phone, password, confirmPassword, country]
      properties:
        username:
          type: string
          minLength: 3
        phone:
          type: string
        password:
          type: string
          minLength: 6
        confirmPassword:
          type: string
        country:
          type: string
          description: Country chosen by the user at registration
        referralCode:
          type: string
        deviceId:
          type: string
          description: Unique device identifier for one-account-per-device enforcement
    GoogleAuthRequest:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
          description: Google OAuth2 access token
        deviceId:
          type: string
          description: Unique device identifier for one-account-per-device enforcement
        country:
          type: string
        referralCode:
          type: string
...
    LoginRequest:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Phone number or username
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        token:
          type: string

    # ─── User ──────────────────────────────────────────────────
    UserProfile:
      type: object
      required: [id, username, userId, referralCode, referralCount, isVip, isAdmin, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        phone:
          type: string
          nullable: true
        userId:
          type: string
          description: Display user ID (e.g. MS-00123)
        referralCode:
          type: string
        referralCount:
          type: integer
        isVip:
          type: boolean
        isAdmin:
          type: boolean
        photoUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Contest:
      type: object
      required: [id, title, description, reward, startDate, endDate, isActive]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        reward:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
    CreateContestRequest:
      type: object
      required: [title, description, reward, startDate, endDate]
      properties:
        title:
          type: string
        description:
          type: string
        reward:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean
          default: true
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: object
...